gallehr/cbam#561: update goods with CN codes >8 digits
